+++ OgreMain/include/OgreMaterialSerializer.h
@@ -249,18 +249,18 @@ namespace Ogre {
         void writeGpuProgramRef(const String& attrib, const GpuProgramPtr& program, const GpuProgramParametersSharedPtr& params);
         void writeGpuPrograms(void);
         void writeGPUProgramParameters(const GpuProgramParametersSharedPtr& params, GpuProgramParameters* defaultParams,
-            const int level = 4, const bool useMainBuffer = true);
+            const unsigned short level = 4, const bool useMainBuffer = true);
 		void writeNamedGpuProgramParameters(const GpuProgramParametersSharedPtr& params, GpuProgramParameters* defaultParams,
-			const int level = 4, const bool useMainBuffer = true);
+			const unsigned short level = 4, const bool useMainBuffer = true);
 		void writeLowLevelGpuProgramParameters(const GpuProgramParametersSharedPtr& params, GpuProgramParameters* defaultParams,
-			const int level = 4, const bool useMainBuffer = true);
+			const unsigned short level = 4, const bool useMainBuffer = true);
 		void writeGpuProgramParameter(
 			const String& commandName, const String& identifier, 
 			const GpuProgramParameters::AutoConstantEntry* autoEntry, 
 			const GpuProgramParameters::AutoConstantEntry* defaultAutoEntry, 
 			bool isFloat, size_t physicalIndex, size_t physicalSize,
 			const GpuProgramParametersSharedPtr& params, GpuProgramParameters* defaultParams,
-			const int level, const bool useMainBuffer);
+			const unsigned short level, const bool useMainBuffer);
 		void writeTextureUnit(const TextureUnitState *pTex);
 		void writeSceneBlendFactor(const SceneBlendFactor c_src, const SceneBlendFactor c_dest, 
 			const SceneBlendFactor a_src, const SceneBlendFactor a_dest);
