+++ OgreMain/src/OgreImageResampler.h
@@ -116,7 +116,7 @@ struct LinearResampler {
 		// for the center of the destination pixel, not the top-left corner
 		uint64 sz_48 = (stepz >> 1) - 1;
 		for (size_t z = dst.front; z < dst.back; z++, sz_48+=stepz) {
-			temp = sz_48 >> 32;
+			temp = static_cast<unsigned int>(sz_48 >> 32);
 			temp = (temp > 0x8000)? temp - 0x8000 : 0;
 			size_t sz1 = temp >> 16;				 // src z, sample #1
 			size_t sz2 = std::min(sz1+1,src.getDepth()-1);// src z, sample #2
@@ -124,7 +124,7 @@ struct LinearResampler {
 
 			uint64 sy_48 = (stepy >> 1) - 1;
 			for (size_t y = dst.top; y < dst.bottom; y++, sy_48+=stepy) {
-				temp = sy_48 >> 32;
+				temp = static_cast<unsigned int>(sy_48 >> 32);
 				temp = (temp > 0x8000)? temp - 0x8000 : 0;
 				size_t sy1 = temp >> 16;					// src y #1
 				size_t sy2 = std::min(sy1+1,src.getHeight()-1);// src y #2
@@ -132,7 +132,7 @@ struct LinearResampler {
 				
 				uint64 sx_48 = (stepx >> 1) - 1;
 				for (size_t x = dst.left; x < dst.right; x++, sx_48+=stepx) {
-					temp = sx_48 >> 32;
+					temp = static_cast<unsigned int>(sx_48 >> 32);
 					temp = (temp > 0x8000)? temp - 0x8000 : 0;
 					size_t sx1 = temp >> 16;					// src x #1
 					size_t sx2 = std::min(sx1+1,src.getWidth()-1);// src x #2
@@ -201,7 +201,7 @@ struct LinearResampler_Float32 {
 		// for the center of the destination pixel, not the top-left corner
 		uint64 sz_48 = (stepz >> 1) - 1;
 		for (size_t z = dst.front; z < dst.back; z++, sz_48+=stepz) {
-			temp = sz_48 >> 32;
+			temp = static_cast<unsigned int>(sz_48 >> 32);
 			temp = (temp > 0x8000)? temp - 0x8000 : 0;
 			size_t sz1 = temp >> 16;				 // src z, sample #1
 			size_t sz2 = std::min(sz1+1,src.getDepth()-1);// src z, sample #2
@@ -209,7 +209,7 @@ struct LinearResampler_Float32 {
 
 			uint64 sy_48 = (stepy >> 1) - 1;
 			for (size_t y = dst.top; y < dst.bottom; y++, sy_48+=stepy) {
-				temp = sy_48 >> 32;
+				temp = static_cast<unsigned int>(sy_48 >> 32);
 				temp = (temp > 0x8000)? temp - 0x8000 : 0;
 				size_t sy1 = temp >> 16;					// src y #1
 				size_t sy2 = std::min(sy1+1,src.getHeight()-1);// src y #2
@@ -217,7 +217,7 @@ struct LinearResampler_Float32 {
 				
 				uint64 sx_48 = (stepx >> 1) - 1;
 				for (size_t x = dst.left; x < dst.right; x++, sx_48+=stepx) {
-					temp = sx_48 >> 32;
+					temp = static_cast<unsigned int>(sx_48 >> 32);
 					temp = (temp > 0x8000)? temp - 0x8000 : 0;
 					size_t sx1 = temp >> 16;					// src x #1
 					size_t sx2 = std::min(sx1+1,src.getWidth()-1);// src x #2
@@ -309,7 +309,7 @@ template<unsigned int channels> struct LinearResampler_Byte {
 		
 		uint64 sy_48 = (stepy >> 1) - 1;
 		for (size_t y = dst.top; y < dst.bottom; y++, sy_48+=stepy) {
-			temp = sy_48 >> 36;
+			temp = static_cast<unsigned int>(sy_48 >> 36);
 			temp = (temp > 0x800)? temp - 0x800: 0;
 			unsigned int syf = temp & 0xFFF;
 			size_t sy1 = temp >> 12;
@@ -319,7 +319,7 @@ template<unsigned int channels> struct LinearResampler_Byte {
 
 			uint64 sx_48 = (stepx >> 1) - 1;
 			for (size_t x = dst.left; x < dst.right; x++, sx_48+=stepx) {
-				temp = sx_48 >> 36;
+				temp = static_cast<unsigned int>(sx_48 >> 36);
 				temp = (temp > 0x800)? temp - 0x800 : 0;
 				unsigned int sxf = temp & 0xFFF;
 				size_t sx1 = temp >> 12;
@@ -334,7 +334,7 @@ template<unsigned int channels> struct LinearResampler_Byte {
 						srcdata[(sx2 + syoff2)*channels+k]*sxfsyf;
 					// accum is computed using 8/24-bit fixed-point math
 					// (maximum is 0xFF000000; rounding will not cause overflow)
-					*pdst++ = (accum + 0x800000) >> 24;
+					*pdst++ = static_cast<uchar>((accum + 0x800000) >> 24);
 				}
 			}
 			pdst += channels*dst.getRowSkip();
